refactor: update TemplateEngine namespace
