+++ src/WeihanLi.Common/Template/DefaultTemplateParser.cs
@@ -3,7 +3,7 @@
 
 using System.Text.RegularExpressions;
 
-namespace WeihanLi.Common.Templating;
+namespace WeihanLi.Common.Template;
 
 internal sealed class DefaultTemplateParser : ITemplateParser
 {
