+++ src/WeihanLi.Common/Template/EnvRenderMiddleware.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Weihan Li. All rights reserved.
 // Licensed under the Apache license.
 
-namespace WeihanLi.Common.Templating;
+namespace WeihanLi.Common.Template;
 
 internal sealed class EnvRenderMiddleware : IRenderMiddleware
 {
