+++ src/WeihanLi.Common/Template/ITemplateParser.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Weihan Li. All rights reserved.
 // Licensed under the Apache license.
 
-namespace WeihanLi.Common.Templating;
+namespace WeihanLi.Common.Template;
 public interface ITemplateParser
 {
     Task<TemplateRenderContext> ParseAsync(string text);
