+++ src/WeihanLi.Common/Template/TemplateEngine.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Weihan Li. All rights reserved.
 // Licensed under the Apache license.
 
-namespace WeihanLi.Common.Templating;
+namespace WeihanLi.Common.Template;
 public sealed class TemplateEngine(ITemplateParser templateParser, ITemplateRenderer templateRenderer)
     : ITemplateEngine, ITemplateRenderer, ITemplateParser
 {
