+++ src/WeihanLi.Common/Template/TemplateRendererBuilder.cs
@@ -3,7 +3,7 @@
 
 using WeihanLi.Common.Helpers;
 
-namespace WeihanLi.Common.Templating;
+namespace WeihanLi.Common.Template;
 
 internal sealed class TemplateEngineBuilder : ITemplateEngineBuilder
 {
